Add workflow to build and package all Qemu artifacts
Add a workflow that builds Qemu statically and collects all necessary
artifacts (qemu data files and binry). Furthermore, use a file under
qemu for arch-specific config for building Qemu. In a similar way
provide a list of files which are produced during the build but are not
necessary as artifacts.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>