Create an action for searching an artifact
Checking if an artifact exists takes palce in various jobs. Therefore,
create an action to check for the existence of an artifact. The artifact
takes as an input the name of the artifact to check and a token. It
returns the name of the artifact, a boolean value that specifies if it
exists or not and if it exists the run_id of the workflow that generated
it and the download url. The action only searches for non-expired artifacts
and returns only the first one that it finds.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>