Refactor build workflows to include searching for existing artifacts
Change the way the build and release workflows operate and move the
functionality for checking if an artifact exists inside the build
workflows. In that way the build workflows can return a unified output
either if the artifact existed or was built contianing the run_id and
the artifact name. Then, the caller can easily download these specific
artifacts with the name and run_id information.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>